legal-navigator/requirements.txt
google-generativeai
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv
